+++ R/leaflet.extras2-package.R
@@ -7,6 +7,7 @@
 #' @import leaflet
 #' @importFrom htmltools htmlDependency tagGetAttribute tags tagList
 #' @importFrom utils globalVariables adist packageVersion
+#' @importFrom grDevices colorRampPalette
 ## usethis namespace: end
 NULL